create randomvariable module
