+++ pyrenew/metaclass.py
@@ -5,22 +5,18 @@ pyrenew helper classes
 """
 
 from abc import ABCMeta, abstractmethod
-from typing import Callable, NamedTuple, Self, get_type_hints
+from typing import NamedTuple, get_type_hints
 
 import jax
 import jax.numpy as jnp
 import jax.random as jr
 import matplotlib.pyplot as plt
 import numpy as np
-import numpyro
-import numpyro.distributions as dist
 import polars as pl
 from jax.typing import ArrayLike
 from numpyro.infer import MCMC, NUTS, Predictive
-from numpyro.infer.reparam import Reparam
 
 from pyrenew.mcmcutils import plot_posterior, spread_draws
-from pyrenew.transformation import Transform
 
 
 def _assert_type(arg_name: str, value, expected_type) -> None:
@@ -309,338 +305,6 @@ class RandomVariable(metaclass=ABCMeta):
         return self.sample(**kwargs)
 
 
-class DynamicDistributionalRV(RandomVariable):
-    """
-    Wrapper class for random variables that sample
-    from a single :class:`numpyro.distributions.Distribution`
-    that is parameterized / instantiated at `sample()` time
-    (rather than at RandomVariable instantiation time).
-    """
-
-    def __init__(
-        self,
-        name: str,
-        distribution_constructor: Callable,
-        reparam: Reparam = None,
-        expand_by_shape: tuple = None,
-    ) -> None:
-        """
-        Default constructor for DynamicDistributionalRV.
-
-        Parameters
-        ----------
-        name : str
-            Name of the random variable.
-        distribution_constructor : Callable
-            Callable that returns a concrete parametrized
-            numpyro.Distributions.distribution instance.
-        reparam : numpyro.infer.reparam.Reparam
-            If not None, reparameterize sampling
-            from the distribution according to the
-            given numpyro reparameterizer
-        expand_by_shape : tuple, optional
-            If not None, call :meth:`expand_by()` on the
-            underlying distribution once it is instianted
-            with the given `expand_by_shape`.
-            Default None.
-
-        Returns
-        -------
-        None
-        """
-
-        self.name = name
-        self.validate(distribution_constructor)
-        self.distribution_constructor = distribution_constructor
-        if reparam is not None:
-            self.reparam_dict = {self.name: reparam}
-        else:
-            self.reparam_dict = {}
-        if not (expand_by_shape is None or isinstance(expand_by_shape, tuple)):
-            raise ValueError(
-                "expand_by_shape must be a tuple or be None ",
-                f"Got {type(expand_by_shape)}",
-            )
-        self.expand_by_shape = expand_by_shape
-
-        return None
-
-    @staticmethod
-    def validate(distribution_constructor: any) -> None:
-        """
-        Confirm that the distribution_constructor is
-        callable.
-
-        Parameters
-        ----------
-        distribution_constructor : any
-            Putative distribution_constructor to validate.
-
-        Returns
-        -------
-        None or raises a ValueError
-        """
-        if not callable(distribution_constructor):
-            raise ValueError(
-                "To instantiate a DynamicDistributionalRV, ",
-                "one must provide a Callable that returns a "
-                "numpyro.distributions.Distribution as the "
-                "distribution_constructor argument. "
-                f"Got {type(distribution_constructor)}, which "
-                "does not appear to be callable",
-            )
-        return None
-
-    def sample(
-        self,
-        *args,
-        obs: ArrayLike = None,
-        **kwargs,
-    ) -> tuple:
-        """
-        Sample from the distributional rv.
-
-        Parameters
-        ----------
-        *args :
-            Positional arguments passed to self.distribution_constructor
-        obs : ArrayLike, optional
-            Observations passed as the `obs` argument to
-            :meth:`numpyro.sample()`. Default `None`.
-        **kwargs : dict, optional
-            Keyword arguments passed to self.distribution_constructor
-
-        Returns
-        -------
-        SampledValue
-           Containing a sample from the distribution.
-        """
-        distribution = self.distribution_constructor(*args, **kwargs)
-        if self.expand_by_shape is not None:
-            distribution = distribution.expand_by(self.expand_by_shape)
-        with numpyro.handlers.reparam(config=self.reparam_dict):
-            sample = numpyro.sample(
-                name=self.name,
-                fn=distribution,
-                obs=obs,
-            )
-        return (
-            SampledValue(
-                sample,
-                t_start=self.t_start,
-                t_unit=self.t_unit,
-            ),
-        )
-
-    def expand_by(self, sample_shape) -> Self:
-        """
-        Expand the distribution by a given
-        shape_shape, if possible. Returns a
-        new DynamicDistributionalRV whose underlying
-        distribution will be expanded by the given shape
-        at sample() time.
-
-        Parameters
-        ----------
-        sample_shape : tuple
-            Sample shape by which to expand the distribution.
-            Passed to the expand_by() method of
-            :class:`numpyro.distributions.Distribution`
-            after the distribution is instantiated.
-
-        Returns
-        -------
-        DynamicDistributionalRV
-            Whose underlying distribution will be expanded by
-            the given sample shape at sampling time.
-        """
-        return DynamicDistributionalRV(
-            name=self.name,
-            distribution_constructor=self.distribution_constructor,
-            reparam=self.reparam_dict.get(self.name, None),
-            expand_by_shape=sample_shape,
-        )
-
-
-class StaticDistributionalRV(RandomVariable):
-    """
-    Wrapper class for random variables that sample
-    from a single :class:`numpyro.distributions.Distribution`
-    that is parameterized / instantiated at RandomVariable
-    instantiation time (rather than at `sample()`-ing time).
-    """
-
-    def __init__(
-        self,
-        name: str,
-        distribution: numpyro.distributions.Distribution,
-        reparam: Reparam = None,
-    ) -> None:
-        """
-        Default constructor for DistributionalRV.
-
-        Parameters
-        ----------
-        name : str
-            Name of the random variable.
-        distribution : numpyro.distributions.Distribution
-            Distribution of the random variable.
-        reparam : numpyro.infer.reparam.Reparam
-            If not None, reparameterize sampling
-            from the distribution according to the
-            given numpyro reparameterizer
-
-        Returns
-        -------
-        None
-        """
-
-        self.name = name
-        self.validate(distribution)
-        self.distribution = distribution
-        if reparam is not None:
-            self.reparam_dict = {self.name: reparam}
-        else:
-            self.reparam_dict = {}
-
-        return None
-
-    @staticmethod
-    def validate(distribution: any) -> None:
-        """
-        Validation of the distribution.
-        """
-        if not isinstance(distribution, numpyro.distributions.Distribution):
-            raise ValueError(
-                "distribution should be an instance of "
-                "numpyro.distributions.Distribution, got "
-                "{type(distribution)}"
-            )
-
-        return None
-
-    def sample(
-        self,
-        obs: ArrayLike | None = None,
-        **kwargs,
-    ) -> tuple:
-        """
-        Sample from the distribution.
-
-        Parameters
-        ----------
-        obs : ArrayLike, optional
-            Observations passed as the `obs` argument to
-            :meth:`numpyro.sample()`. Default `None`.
-        **kwargs : dict, optional
-            Additional keyword arguments passed through
-            to internal sample calls, should there be any.
-
-        Returns
-        -------
-        SampledValue
-           Containing a sample from the distribution.
-        """
-        with numpyro.handlers.reparam(config=self.reparam_dict):
-            sample = numpyro.sample(
-                name=self.name,
-                fn=self.distribution,
-                obs=obs,
-            )
-        return (
-            SampledValue(
-                sample,
-                t_start=self.t_start,
-                t_unit=self.t_unit,
-            ),
-        )
-
-    def expand_by(self, sample_shape) -> Self:
-        """
-        Expand the distribution by the given sample_shape,
-        if possible. Returns a new StaticDistributionalRV
-        whose underlying distribution has been expanded by
-        the given sample_shape via
-        :meth:`~numpyro.distributions.Distribution.expand_by()`
-
-        Parameters
-        ----------
-        sample_shape : tuple
-            Sample shape for the expansion. Passed to the
-            :meth:`expand_by()` method of
-            :class:`numpyro.distributions.Distribution`.
-
-        Returns
-        -------
-        StaticDistributionalRV
-            Whose underlying distribution has been expanded by
-            the given sample shape.
-        """
-        if not isinstance(sample_shape, tuple):
-            raise ValueError(
-                "sample_shape for expand()-ing "
-                "a DistributionalRV must be a "
-                f"tuple. Got {type(sample_shape)}"
-            )
-        return StaticDistributionalRV(
-            name=self.name,
-            distribution=self.distribution.expand_by(sample_shape),
-            reparam=self.reparam_dict.get(self.name, None),
-        )
-
-
-def DistributionalRV(
-    name: str,
-    distribution: numpyro.distributions.Distribution | Callable,
-    reparam: Reparam = None,
-) -> RandomVariable:
-    """
-    Factory function to generate Distributional RandomVariables,
-    either static or dynamic.
-
-    Parameters
-    ----------
-    name : str
-        Name of the random variable.
-
-    distribution: numpyro.distributions.Distribution | Callable
-        Either numpyro.distributions.Distribution instance
-        given the static distribution of the random variable or
-        a callable that returns a parameterized
-        numpyro.distributions.Distribution when called, which
-        allows for dynamically-parameterized DistributionalRVs,
-        e.g. a Normal distribution with an inferred location and
-        scale.
-
-    reparam : numpyro.infer.reparam.Reparam
-        If not None, reparameterize sampling
-        from the distribution according to the
-        given numpyro reparameterizer
-
-    Returns
-    -------
-    DynamicDistributionalRV | StaticDistributionalRV or
-    raises a ValueError if a distribution cannot be constructed.
-    """
-    if isinstance(distribution, dist.Distribution):
-        return StaticDistributionalRV(
-            name=name, distribution=distribution, reparam=reparam
-        )
-    elif callable(distribution):
-        return DynamicDistributionalRV(
-            name=name, distribution_constructor=distribution, reparam=reparam
-        )
-    else:
-        raise ValueError(
-            "distribution argument to DistributionalRV "
-            "must be either a numpyro.distributions.Distribution "
-            "(for instantiating a static DistributionalRV) "
-            "or a callable that returns a "
-            "numpyro.distributions.Distribution (for "
-            "a dynamic DistributionalRV"
-        )
-
-
 class Model(metaclass=ABCMeta):
     """Abstract base class for models"""
 
@@ -924,137 +588,3 @@ class Model(metaclass=ABCMeta):
         )
 
         return predictive(rng_key, **kwargs)
-
-
-class TransformedRandomVariable(RandomVariable):
-    """
-    Class to represent RandomVariables defined
-    by taking the output of another RV's
-    :meth:`RandomVariable.sample()` method
-    and transforming it by a given transformation
-    (typically a :class:`Transform`)
-    """
-
-    def __init__(
-        self,
-        name: str,
-        base_rv: RandomVariable,
-        transforms: Transform | tuple[Transform],
-    ):
-        """
-        Default constructor
-
-        Parameters
-        ----------
-        name : str
-            A name for the random variable instance.
-        base_rv : RandomVariable
-            The underlying (untransformed) RandomVariable.
-        transforms : Transform
-            Transformation or tuple of transformations
-            to apply to the output of
-            `base_rv.sample()`; single values will be coerced to
-            a length-one tuple. If a tuple, should be the same
-            length as the tuple returned by `base_rv.sample()`.
-
-        Returns
-        -------
-        None
-        """
-        self.name = name
-        self.base_rv = base_rv
-
-        if not isinstance(transforms, tuple):
-            transforms = (transforms,)
-        self.transforms = transforms
-        self.validate()
-
-    def sample(self, record=False, **kwargs) -> tuple:
-        """
-        Sample method. Call self.base_rv.sample()
-        and then apply the transforms specified
-        in self.transforms.
-
-        Parameters
-        ----------
-        record : bool, optional
-            Whether to record the value of the deterministic
-            RandomVariable. Defaults to False.
-        **kwargs :
-            Keyword arguments passed to self.base_rv.sample()
-
-        Returns
-        -------
-        tuple of the same length as the tuple returned by
-        self.base_rv.sample()
-        """
-
-        untransformed_values = self.base_rv.sample(**kwargs)
-        transformed_values = tuple(
-            SampledValue(
-                t(uv.value),
-                t_start=self.t_start,
-                t_unit=self.t_unit,
-            )
-            for t, uv in zip(self.transforms, untransformed_values)
-        )
-
-        if record:
-            if len(untransformed_values) == 1:
-                numpyro.deterministic(self.name, transformed_values[0].value)
-            else:
-                suffixes = (
-                    untransformed_values._fields
-                    if hasattr(untransformed_values, "_fields")
-                    else range(len(transformed_values))
-                )
-                for suffix, tv in zip(suffixes, transformed_values):
-                    numpyro.deterministic(f"{self.name}_{suffix}", tv.value)
-
-        return transformed_values
-
-    def sample_length(self):
-        """
-        Sample length for a transformed
-        random variable must be equal to the
-        length of self.transforms or
-        validation will fail.
-
-        Returns
-        -------
-        int
-           Equal to the length self.transforms
-        """
-        return len(self.transforms)
-
-    def validate(self):
-        """
-        Perform validation checks on a
-        TransformedRandomVariable instance,
-        confirming that all transformations
-        are callable and that the number of
-        transformations is equal to the sample
-        length of the base random variable.
-
-        Returns
-        -------
-        None
-           on successful validation, or raise a ValueError
-        """
-        for t in self.transforms:
-            if not callable(t):
-                raise ValueError(
-                    "All entries in self.transforms " "must be callable"
-                )
-        if hasattr(self.base_rv, "sample_length"):
-            n_transforms = len(self.transforms)
-            n_entries = self.base_rv.sample_length()
-            if not n_transforms == n_entries:
-                raise ValueError(
-                    "There must be exactly as many transformations "
-                    "specified as entries self.transforms as there are "
-                    "entries in the tuple returned by "
-                    "self.base_rv.sample()."
-                    f"Got {n_transforms} transforms and {n_entries} "
-                    "entries"
-                )
